Fix crashes...

+++ src/player/PlayerAPI.php
@@ -73,7 +73,7 @@ class PlayerAPI implements IPlayerAPI {
 	private float $slimeBlockTicks = 0.0;
 	private float $deathTicks = 0.0;
 	private float $placingTicks = 0.0;
-	private int $lastMoveTick = 0;
+	private float $lastMoveTick = 0.0;
 	private int $cps = 0;
 	private int $blocksBrokeASec = 0;
 	private int $blocksPlacedASec = 0;
@@ -174,11 +174,11 @@ class PlayerAPI implements IPlayerAPI {
 		$this->onPlant = $data;
 	}
 
-	public function setLastMoveTick(int $data) : void {
+	public function setLastMoveTick(float $data) : void {
 		$this->lastMoveTick = $data;
 	}
 
-	public function getLastMoveTick() : int {
+	public function getLastMoveTick() : float {
 		return $this->lastMoveTick;
 	}
